Create outlook_to_pdf.ahk
The actual code with which I helped a colleague get around their version of Outlook not having a way to mass save emails with days left until inbox shutdown. Sometime it messes up but only after hundreds of emails are saved. You only have to change file naming, incrementing "batch" by +1 (otherwise emails on next run will overwrite emails from previous run).

This code is very sensitive to how fast or slow PC is, better to do this while plugged in (yes computer runs slower in a way that's not humanly noticeable, but the code does suffer this slowdown, one can play around with all the Sleep commands to get it to work just right).

This is the first thing I made on AutoHotKeys. Please steal and make better!